renamed files to correct views
