+++ client/screens/Shows.js
@@ -3,12 +3,12 @@ import { StyleSheet, Text, View } from 'react-native';
 
 import MenuButton from '../components/MenuButton'
 
-export default class SettingsScreen extends React.Component {
+export default class HomeScreen extends React.Component {
   render() {
     return (
       <View style={styles.container}>
-        <MenuButton navigation={this.props.navigation} />
-        <Text style={styles.text}>Settings</Text>
+        <MenuButton navigation={this.props.navigation}/>
+        <Text style={styles.text}>Shows</Text>
       </View>
     )
   }
